Fix issue with boss splits
If a new run was started with a defeated boss still in the enemy table, it would split immediately after starting a run. It now waits to enable the boss splits until the first battle of the game is detected, when we know the enemy HP table has been updated.